Альтернативный сетевой слой и вёрстка экрана со списком nft колекций

+++ FakeNFT/Scenes /Profile/MyNFT/MyNFTTableCell.swift
@@ -96,7 +96,7 @@ final class MyNFTTableCell: UITableViewCell {
     func configure(
         nftImageURL: URL?,
         nftTitle: String,
-        nftPrice: Float,
+        nftPrice: Double,
         nftAuthor: String,
         nftRatingStars: Int,
         nftIsLiked: Bool,
